New User Message Policy and feature test
